requirements.txt
feedparser>=6.0.10
requests>=2.31.0
PyYAML>=6.0.1
Jinja2>=3.1.2
beautifulsoup4>=4.12.0
